ref: move sentry.runner.__init__.main to sentry.runner.main.main
first step in breaking some nasty import cycles in sentry.runner.*
