+++ setup.cfg
@@ -32,7 +32,7 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    sentry = sentry.runner:main
+    sentry = sentry.__main__:main
 sentry.apps =
     # TODO: This can be removed once the getsentry tests no longer check for this app
     auth_activedirectory = sentry.auth.providers.saml2.activedirectory
