+++ src/sentry/__main__.py
@@ -1,4 +1,4 @@
-from .runner import main
+from .runner.main import main
 
 if __name__ == "__main__":
     main()
